rename drift_detector.py to detector.py
