+++ README.md
@@ -108,7 +108,7 @@ Menelaus should work with Python 3.8 or higher.
 
 # Getting Started
 
-Each detector implements the API defined by `menelaus.drift_detector`:
+Each detector implements the API defined by `menelaus.detector`:
 notably, they have an `update` method which allows new data to be passed, and a `drift_state` attribute which tells the user whether drift has been
 detected, along with (usually) other attributes specific to the detector class.
