+++ menelaus/concept_drift/ddm.py
@@ -1,5 +1,5 @@
 import numpy as np
-from menelaus.drift_detector import StreamingDetector
+from menelaus.detector import StreamingDetector
 
 
 class DDM(StreamingDetector):
