+++ menelaus/concept_drift/eddm.py
@@ -1,5 +1,5 @@
 import numpy as np
-from menelaus.drift_detector import StreamingDetector
+from menelaus.detector import StreamingDetector
 
 
 class EDDM(StreamingDetector):
