+++ menelaus/detector.py
